+++ executor-service/src/main/java/executor/service/exception/stepexception/ClickXPathException.java
@@ -1,4 +1,5 @@
-package executor.service.exception;
+package executor.service.exception.stepexception;
+
 
 public class ClickXPathException extends StepExecutionException {
     public ClickXPathException(Throwable cause) {
