+++ executor-service/src/main/java/executor/service/exception/stepexception/StepExecutionException.java
@@ -1,4 +1,4 @@
-package executor.service.exception;
+package executor.service.exception.stepexception;
 
 public class StepExecutionException extends RuntimeException {
     public StepExecutionException(String message, Throwable cause) {
